% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_losses.R
\name{damage_per_loc}
\alias{damage_per_loc}
\title{Economic impact of temperature}
\usage{
damage_per_loc(temp_1_locs, temp_2_locs, econ_scores)
}
\arguments{
\item{temp_1_locs:}{a list of locations experiencing temp #1.}

\item{temp_2_locs:}{a list of locations experiencing temp #2.}

\item{econ_scores:}{a dataframe containing columns for each location ('Location') and their associated economic scores ("Scores') from 1 to 5.}

\item{df_temps1:}{a dataframe containing the locations experiencing temp #1 and their associated economic losses.}

\item{df_temps2:}{a dataframe containing the locations experiencing temp #2 and their associated economic losses.}

\item{tot_damage:}{total economic losses across all locations due to both temperatures.}
}
\description{
This function returns the economic damages associated with temp 1 and temp 2 per location, as well as the total economic losses associated with both temperatures across all locations.
}
\note{
Economic scores represent damage brackets (in USD) as defined below...
1 =  $1,400,000
2 = $7,100,000
3 = $30,000,000
4 = $44,000,000
5 = $600,000,000.
}
