% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_water_temps.R
\name{Max_2_temps}
\alias{Max_2_temps}
\title{Top two warmest water temperatures across locations}
\usage{
Max_2_temps(temps, locs, loc_sizes)
}
\arguments{
\item{df_raw:}{a dataframe with columns for locations ('locs'), temperatures ('temps'), sizes of each location in km ('loc_sizes')}
}
\value{
The warmest and second warmest temperatures, a list of all the locations experiencing the warmest and 2nd warmest temperatures, the associated sizes with each locaton, and the total area experiencing the warmest temperature and the 2nd warmest temperature.
}
\description{
This function returns the top two temperatures experienced within a given set of locations, a list of the locations experiencing these two temperatures, and the total geographic area experiencing these two temperatures in km.
}
