% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_cap_per_km_losses.R
\name{temp_damages_km_cap}
\alias{temp_damages_km_cap}
\title{Per km and per capita economic impacts of temperature}
\usage{
temp_damages_km_cap(df_temps1, df_temps2, pops_sizes)
}
\arguments{
\item{df_temps1:}{A dataframe with columns for economic losses ("Scores") and locations ("Location") for temp 1}

\item{pops_sizes:}{A dataframe with columns for population sizes ("population"), location sizes in km ("km"), and locations ("Location").}

\item{df_temps2:}{A dataframe with columns for economic losses ("Scores") and locations ("Location") for temp 2.}
}
\value{
One dataframe for temp 1 and another for temp 2, each with columns for locations and the per capita economic losses for each location. Also, the per km economic losses due to temp #1, and the per km economic losses due to temp #2.
}
\description{
This function provides the per km economic losses for a list of locations associated with two temperatures.
}
\author{
Kat Leigh & Anna Ableman
}
