% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_sizes.R
\docType{data}
\name{pops_sizes}
\alias{pops_sizes}
\title{Population Size and Location Data Frame}
\format{Individual vectors with list of each variable were initially created. A data frame was created in tidy format.}
\source{
This data was randomly created
\describe{
\population - a list of the population sizes for each location
\locs - a list of 10 different California locations
\loc_sizes - a list of the sizes in kilometers per location
\pops_sizes - a data frame with the 10 locations and the corresponding population sizes and the location sizes (km)
}
}
\usage{
pops_sizes
}
\description{
A sample data frames created that includes vectors of: population, location, and size of location in kilometers.
}
\keyword{datasets}
